Add files via upload
Dola Protocol is the core architecture of OmniLending, which is built on Sui Network.

It uses "Hub and Spoke" model to realize the omnichain liquidity. All of the new spoke chains can easily plug and play into Dola Protocol.

Via Dola Protocol, OmniBTC can easily connect all the on chain liquidity and realize an omnichain ecosystem.